// src/app/model/user.model.ts
export class User {

  id: number;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  conPassword: string;
  middleName: string;
  phone: string;
}
export class Useredit {

  _id: number;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  conPassword: string;
  middleName: string;
  phone: string;
}